Add options for salmon_index kmer sizes
This PR adds parameters to allow the user to specify the kmer used when building a salmon index. 
I added a `kmer` parameter to each of the `salmon_index`, salmon_sa_index`, and `salmon_partial_sa_index` recipes, with the default a 31, as for `salmon` in general usage. The `salmon_sa_index`, and `salmon_partial_sa_index` recipes did not previously use a -k option, so I have added it there, and standardized the `salmon index` parameter order across all samples.

I believe that I have followed the conventions and expected behaviors of the recipe files, but please let me know if there are any changes that might be required.